+++ CHANGELOG.md
@@ -5,6 +5,7 @@ See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 ## 0.4.0 - 2023-09-02
 
 * Added
+  * PnP editor: after manual selection, the component is appliet to all items where Comment and Footprint matches the current item
 * Changed
   * DB components editor redesigned - now it's able to handle even 10'000 elements
 * Deprecated